refactor: extracted survey data types for tRPC
