+++ src/panels/DocumentTab.ts
@@ -9,7 +9,8 @@ import vscode from 'vscode';
 import { type NoSqlQueryConnection } from '../docdb/NoSqlCodeLensProvider';
 import { DocumentSession } from '../docdb/session/DocumentSession';
 import { type CosmosDbRecordIdentifier } from '../docdb/types/queryResult';
-import { ExperienceKind, promptAfterActionEventually, UsageImpact } from '../utils/survey';
+import { promptAfterActionEventually } from '../utils/survey';
+import { ExperienceKind, UsageImpact } from '../utils/surveyTypes';
 import { BaseTab, type CommandPayload } from './BaseTab';
 
 type DocumentTabMode = 'add' | 'edit' | 'view';
