+++ src/panels/QueryEditorTab.ts
@@ -14,7 +14,8 @@ import { DocumentSession } from '../docdb/session/DocumentSession';
 import { QuerySession } from '../docdb/session/QuerySession';
 import { type CosmosDbRecordIdentifier, type ResultViewMetadata } from '../docdb/types/queryResult';
 import { getNoSqlQueryConnection } from '../docdb/utils/NoSqlQueryConnection';
-import { ExperienceKind, promptAfterActionEventually, UsageImpact } from '../utils/survey';
+import { promptAfterActionEventually } from '../utils/survey';
+import { ExperienceKind, UsageImpact } from '../utils/surveyTypes';
 import * as vscodeUtil from '../utils/vscodeUtils';
 import { BaseTab, type CommandPayload } from './BaseTab';
 import { DocumentTab } from './DocumentTab';
