+++ src/tree/docdb/DocumentDBDatabaseResourceItem.ts
@@ -8,7 +8,8 @@ import { createContextValue } from '@microsoft/vscode-azext-utils';
 import vscode, { type TreeItem } from 'vscode';
 import { type Experience } from '../../AzureDBExperiences';
 import { getCosmosClient } from '../../docdb/getCosmosClient';
-import { countExperienceUsageForSurvey, ExperienceKind, UsageImpact } from '../../utils/survey';
+import { countExperienceUsageForSurvey } from '../../utils/survey';
+import { ExperienceKind, UsageImpact } from '../../utils/surveyTypes';
 import { type CosmosDBTreeElement } from '../CosmosDBTreeElement';
 import { type TreeElementWithContextValue } from '../TreeElementWithContextValue';
 import { type TreeElementWithExperience } from '../TreeElementWithExperience';
