+++ src/tree/docdb/DocumentDBItemsResourceItem.ts
@@ -8,7 +8,8 @@ import { createContextValue, createGenericElement, type IActionContext } from '@
 import vscode, { type TreeItem } from 'vscode';
 import { type Experience } from '../../AzureDBExperiences';
 import { getCosmosClient } from '../../docdb/getCosmosClient';
-import { countExperienceUsageForSurvey, ExperienceKind, UsageImpact } from '../../utils/survey';
+import { countExperienceUsageForSurvey } from '../../utils/survey';
+import { ExperienceKind, UsageImpact } from '../../utils/surveyTypes';
 import { getBatchSizeSetting } from '../../utils/workspacUtils';
 import { type CosmosDBTreeElement } from '../CosmosDBTreeElement';
 import { type TreeElementWithContextValue } from '../TreeElementWithContextValue';
