+++ src/utils/surveyTypes.ts
@@ -1,0 +1,15 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+export enum UsageImpact {
+    Low = 5,
+    Medium = 10,
+    High = 20,
+}
+
+export enum ExperienceKind {
+    Mongo = 'Mongo',
+    NoSQL = 'NoSQL',
+}
